Add files via upload
1)Добавлен вывод скорости шарика на экран
2)Добавлен выход из программы с помощью кнопки 
3)Еще какие-то вещи, но я забыл....
