+++ Header.hpp
@@ -1,0 +1,13 @@
+//
+// Created by behap on 21.03.2021.
+//
+
+#ifndef PHYSIC_HEADER_HPP
+#define PHYSIC_HEADER_HPP
+
+#include "Cannon.hpp"
+#include "Block.hpp"
+#include "Land.hpp"
+#include "numbers.h"
+
+#endif //PHYSIC_HEADER_HPP
