+++ images/Ball.png.url
@@ -1,0 +1,2 @@
+[InternetShortcut]
+URL=https://github.com/constadry/physic-model/blob/main/images/Ball.png
